requirements.txt
# AutoCI Modified Godot Requirements

# Core Dependencies
# Note: Godot engine is external and uses C# for scripting
# panda3d removed - using modified Godot instead
numpy>=1.21.0
pillow>=9.0.0

# AI & Machine Learning
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
safetensors>=0.3.1

# Automation & Control
pyautogui>=0.9.53
keyboard>=0.13.5
mouse>=0.7.1
python-xlib>=0.31; sys_platform == 'linux'
pywin32>=305; sys_platform == 'win32'

# Web & Networking
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio[asyncio]>=5.9.0
python-engineio>=4.7.0
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=0.8.0
websockets>=11.0

# Korean NLP (Optional)
konlpy>=0.6.0
jamo>=0.4.1

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.14.0

# System Monitoring
psutil>=5.9.0
gputil>=1.4.0
py-cpuinfo>=9.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Development Tools
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.2
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.3.0
click>=8.1.3
typer>=0.9.0
fire>=0.5.0
questionary>=1.10.0
prompt-toolkit>=3.0.38

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
markdown>=3.4.0
pygments>=2.15.0

# Logging
loguru>=0.7.0
sentry-sdk>=1.25.0

# Game Development Specific
# pygame>=2.4.0  # Not needed - Godot handles graphics/sound
# pymunk>=6.4.0  # Not needed - Godot has built-in physics
noise>=1.2.2   # Procedural generation
pydub>=0.25.1  # Audio processing

# C# Integration
pythonnet>=3.0.1  # Python.NET for C# interop

# Image Processing
opencv-python>=4.8.0

# Web Scraping (for information gathering)
beautifulsoup4>=4.12.0
selenium>=4.10.0
googlesearch-python>=1.2.3

# Security
cryptography>=41.0.0
pyjwt>=2.7.0

# Additional Dependencies
python-dateutil>=2.8.2
pytz>=2023.3
ujson>=5.8.0
screeninfo>=0.8.1
pynput>=1.7.6